Move the JSON Patch helper into imbi-common
imbi-scheduler's PATCH /tasks/{slug} takes RFC 6902 like the rest of the
platform, and the helper that applies one was API-local. It has no
imbi-api coupling -- fastapi, jsonpatch, pydantic -- so it moves rather
than gets copied a second time.

Rewrote the 19 importing modules instead of leaving a re-export shim:
the module name and call sites are unchanged, so the import line is the
only difference and an indirection would earn nothing. Its 13 tests
moved with it.

Verified: 1126 common and 2494 api tests pass, coverage 91.29 and 85.97
percent, lint and basedpyright clean.
