+++ apps/api/src/imbi/api/endpoints/document_templates.py
@@ -13,11 +13,11 @@ import typing
 import fastapi
 import pydantic
 
-from imbi.api import patch as json_patch
 from imbi.api.auth import permissions
 from imbi.api.endpoints._helpers import fetch_or_404
 from imbi.api.graph_sql import props_template, set_clause
 from imbi.common import graph
+from imbi.common import patch as json_patch
 
 LOGGER = logging.getLogger(__name__)
