+++ apps/api/src/imbi/api/endpoints/environments.py
@@ -7,13 +7,13 @@ import typing
 import fastapi
 import pydantic
 
-from imbi.api import patch as json_patch
 from imbi.api.auth import permissions
 from imbi.api.endpoints import plugin_edges as _plugin_edges
 from imbi.api.endpoints._helpers import conflict_on_unique_violation
 from imbi.api.graph_sql import props_template, set_clause
 from imbi.api.relationships import relationship_link
 from imbi.common import blueprints, graph, models
+from imbi.common import patch as json_patch
 
 LOGGER = logging.getLogger(__name__)
