+++ apps/api/src/imbi/api/endpoints/project_types.py
@@ -8,12 +8,12 @@ import fastapi
 import pydantic
 
 from imbi.api import blueprint_attributes
-from imbi.api import patch as json_patch
 from imbi.api.auth import permissions
 from imbi.api.endpoints._helpers import conflict_on_unique_violation
 from imbi.api.graph_sql import props_template, set_clause
 from imbi.api.relationships import relationship_link
 from imbi.common import blueprints, graph, models
+from imbi.common import patch as json_patch
 
 LOGGER = logging.getLogger(__name__)
