+++ apps/api/src/imbi/api/endpoints/tags.py
@@ -15,11 +15,11 @@ import nanoid
 import pydantic
 import slugify
 
-from imbi.api import patch as json_patch
 from imbi.api.auth import permissions
 from imbi.api.endpoints._helpers import conflict_on_unique_violation
 from imbi.api.relationships import RelationshipSpec, build_relationships
 from imbi.common import graph, models
+from imbi.common import patch as json_patch
 
 LOGGER = logging.getLogger(__name__)
